chore: tools dir, one class per file
